app/requirements.txt
Flask==2.2.5
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.31.0

google-cloud-storage==2.10.0
google-cloud-firestore==2.9.0
google-cloud-bigquery==3.12.0
google-auth==2.23.0
